api_examples/requirements.txt
odoorpc==0.9.0
requests==2.31.0
flask==2.3.3
python-dotenv==1.1.1
